# chore: remove defaultOnBanReach helper fn