--- a/index.js
+++ b/index.js
@@ -32,6 +32,8 @@ async function fastifyRateLimit (fastify, settings) {
   if (typeof settings.enableDraftSpec === 'boolean' && settings.enableDraftSpec) {
     globalParams.enableDraftSpec = true
     labels = draftSpecHeaders
+  } else {
+    globalParams.enableDraftSpec = false
   }
 
   globalParams.addHeaders = Object.assign({
@@ -66,12 +68,7 @@ async function fastifyRateLimit (fastify, settings) {
   globalParams.hook = settings.hook || defaultHook
   globalParams.allowList = settings.allowList || settings.whitelist || null
   globalParams.ban = settings.ban || null
-  globalParams.onBanReach = defaultOnBanReach
-
-  if (typeof settings.onBanReach === 'function') {
-    globalParams.onBanReach = settings.onBanReach
-  }
-
+  globalParams.onBanReach = typeof settings.onBanReach === 'function' ? settings.onBanReach : undefined
   globalParams.continueExceeding = settings.continueExceeding || false
 
   // define the name of the app component. Related to redis, it will be use as a part of the keyname define in redis.
@@ -94,9 +91,6 @@ async function fastifyRateLimit (fastify, settings) {
     ? settings.keyGenerator
     : (req) => req.ip
 
-  globalParams.errorResponseBuilder = defaultErrorResponse
-  globalParams.isCustomErrorMessage = false
-
   globalParams.onExceeded = settings.onExceeded
   globalParams.onExceeding = settings.onExceeding
 
@@ -104,13 +98,17 @@ async function fastifyRateLimit (fastify, settings) {
   if (typeof settings.errorResponseBuilder === 'function') {
     globalParams.errorResponseBuilder = settings.errorResponseBuilder
     globalParams.isCustomErrorMessage = true
+  } else {
+    globalParams.errorResponseBuilder = defaultErrorResponse
+    globalParams.isCustomErrorMessage = false
   }
 
-  globalParams.skipOnError = settings.skipOnError || false
+  globalParams.skipOnError = typeof settings.skipOnError === 'boolean' ? settings.skipOnError : false
 
   const run = Symbol('rate-limit-did-run')
   pluginComponent.run = run
   fastify.decorateRequest(run, false)
+
   if (!fastify.hasDecorator('rateLimit')) {
     // The rate limit plugin can be registered multiple times but decorate throws if called multiple times for the same field
     fastify.decorate('rateLimit', function rateLimit (options) {
@@ -265,7 +263,7 @@ function rateLimitRequestHandler (params, pluginComponent) {
 
     if (code === 403) {
       respCtx.ban = true
-      params.onBanReach(req, key)
+      if (typeof params.onBanReach === 'function') params.onBanReach(req, key)
     }
 
     throw params.errorResponseBuilder(req, respCtx)
@@ -277,8 +275,6 @@ function defaultErrorResponse (req, context) {
   err.statusCode = context.statusCode
   return err
 }
-
-function defaultOnBanReach (req, key) {}
 
 module.exports = fp(fastifyRateLimit, {
   fastify: '4.x',