@Uzlopak 2023-08-04
@mcollina 

Are you sure? Is noop not faster than optional chaining?