@gurgunday 2023-08-04
I actually forgot about this one, but the main thing is `onExceeding` and `onExceeded` both had `typeof ===` checks while `onBan` called an empty function, which you say is better, which I don't object to

The PR is simply one about consistency, they should either all have empty default functions or be optionally called — like how either all options should be type checked or none

PS: I think the difference is extremely small