@Uzlopak 2023-08-04
What would be more interesting would be a benchmark