@gurgunday 2023-08-04
Let me try that even though it's likely that the difference will be hard to measure, especially between an empty function call and optional chaining